Rewrite how arguments are obtained to prevent writing files when '>' present.
If a line to be completed contained an output redirection, the old method of
parsing arguments would execute bash with the line, redirecting the list of
arguments to the file.  Now the arguments are obtained from a bash function
(found in bin/getopt_complete.sh) and processed individually preventing this
behaviour.

This does break compatibility with older versions of Getopt::Complete.
Documentation updates &c. are required before this version is ready for a
release.